+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - General
   - Improve intialization feedback report messages.
   - Consider `dev` and `test` alias for deps.edn projects as project-specs during classpath lookup. #586
+  - Avoid scanning source-paths twice, as it was being considered as part of external classpath as well.
 
 ## 2021.09.30-15.28.01